Added an option to incorporate the diversity term of the loss into the total loss of the model, self.total_loss (in SocialVRNN.py)
The loss as specified in the article encompasses a diversity term L_div. This is not the case in the present source code implementation of SocialVRNN. Now, a flag can be passed for the user to choose whether to use the original source code implementation of the total loss function (without div term), and that of the publication (with the div term).